changed convolution to use c2c, not r2c and c2r but it dooesnt' work

+++ CudaProjectImageProcessing/ConvolutionOld.cuh
@@ -5,7 +5,7 @@ class ConvolutionCalculator
 public:
 	static void convolution(
 		const uint8_t* image_in,
-		const float* kernel,
+		const double* kernel,
 		const int image_width,
 		const int image_height,
 		const int kernel_width,
